requirements.txt
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
fonttools==4.32.0
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
numpy-stl==2.16.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-utils==3.1.0
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
